Refactored Show & Edit page
