+++ app/javascript/src/components/Blog/Edit/Header.jsx
@@ -1,0 +1,38 @@
+import React from "react";
+
+import { ActionBlock, ActionMenubar, PageHeader } from "components/common";
+import { ExternalLink } from "neetoicons";
+import { Button } from "neetoui";
+import routes from "routes";
+import buildUrl from "utils/buildUrl";
+import withT from "utils/withT";
+
+const Header = ({
+  t,
+  slug,
+  isPublishButtonActive,
+  setIsPublishButtonActive,
+  handleActionSubmit,
+}) => (
+  <PageHeader label={t("blog.editPost")}>
+    <div className="flex items-center gap-3">
+      <Button
+        icon={ExternalLink}
+        style="text"
+        to={buildUrl(routes.blogs.show, { slug })}
+        tooltipProps={{ content: "Preview", position: "top" }}
+      />
+      <ActionBlock
+        {...{
+          isPublishButtonActive,
+          setIsPublishButtonActive,
+          handleActionSubmit,
+        }}
+        shouldShowMenubar
+      />
+      <ActionMenubar {...{ slug }} />
+    </div>
+  </PageHeader>
+);
+
+export default withT(Header);
