+++ app/javascript/src/components/Blog/Edit/index.jsx
@@ -1,37 +1,29 @@
 import React, { useRef, useState } from "react";
 
-import {
-  ActionBlock,
-  ActionMenubar,
-  PageHeader,
-  PageLoader,
-} from "components/common";
-import { useEditPost, useShowPost } from "hooks/reactQuery/usePostsApi";
-import { ExternalLink } from "neetoicons";
-import { Button } from "neetoui";
-import { useTranslation } from "react-i18next";
-import { useHistory, useParams } from "react-router-dom";
+import { PageLoader, PageNotFound } from "components/common";
+import { useEditPost } from "hooks/reactQuery/usePostsApi";
+import useFetchPostWithStatus from "hooks/useFetchPostWithStatus";
+import { useHistory } from "react-router-dom";
 import routes from "routes";
 import buildUrl from "utils/buildUrl";
 
-import EditPostForm from "./Form";
-import { formatCategoriesForSelectInput, getCategoryIds } from "./utils";
+import Header from "./Header";
+
+import EditPostForm from "../Form";
+import { formatCategoriesForSelectInput, getCategoryIds } from "../utils";
 
 const Edit = () => {
   const [isPublishButtonActive, setIsPublishButtonActive] = useState(true);
 
-  const { t } = useTranslation();
-
   const history = useHistory();
 
-  const { slug } = useParams();
-
-  const { data: post = {}, isLoading: isPostLoading } = useShowPost(slug);
+  const { post, isPostLoading, isPostNotFound } = useFetchPostWithStatus();
 
   const formikRef = useRef(null);
 
-  const { mutate: editPost, isLoading: isSubmissionLoading } =
-    useEditPost(slug);
+  const { mutate: editPost, isLoading: isSubmissionLoading } = useEditPost(
+    post?.slug
+  );
 
   const handleFormSubmit = data => {
     const { title, description, slug, categories = [] } = data;
@@ -65,27 +57,20 @@ const Edit = () => {
     return <PageLoader />;
   }
 
+  if (isPostNotFound) {
+    return <PageNotFound />;
+  }
+
   return (
     <div className="h-full py-12 pl-12">
-      <PageHeader label={t("blog.editPost")}>
-        <div className="flex items-center gap-3">
-          <Button
-            icon={ExternalLink}
-            style="text"
-            to={buildUrl(routes.blogs.show, { slug })}
-            tooltipProps={{ content: "Preview", position: "top" }}
-          />
-          <ActionBlock
-            {...{
-              isPublishButtonActive,
-              setIsPublishButtonActive,
-              handleActionSubmit,
-            }}
-            shouldShowMenubar
-          />
-          <ActionMenubar {...{ slug }} />
-        </div>
-      </PageHeader>
+      <Header
+        {...{
+          slug: post?.slug,
+          isPublishButtonActive,
+          setIsPublishButtonActive,
+          handleActionSubmit,
+        }}
+      />
       <div className="h-11/12 container w-11/12 overflow-y-auto rounded-md border p-3 shadow-sm md:p-12">
         <EditPostForm
           {...{
