+++ app/javascript/src/hooks/useFetchPostWithStatus.js
@@ -1,0 +1,19 @@
+import { useParams } from "react-router-dom";
+
+import { useShowPost } from "./reactQuery/usePostsApi";
+
+const useFetchPostWithStatus = () => {
+  const { slug } = useParams();
+
+  const {
+    data: post = {},
+    isLoading: isPostLoading,
+    error,
+  } = useShowPost(slug);
+
+  const isPostNotFound = error?.status === 404;
+
+  return { post, isPostLoading, isPostNotFound };
+};
+
+export default useFetchPostWithStatus;
